openfeign

+++ parent/base/src/main/java/com/thy/backend/parent/base/cons/BasePersistentCons.java
@@ -4,5 +4,5 @@ package com.thy.backend.parent.base.cons;
  * @author root
  */
 public interface BasePersistentCons {
-    int DEFAULT_CREATOR_MODIFIER = -1;
+    long DEFAULT_CREATOR_MODIFIER = -1L;
 }
